more apis

+++ src/main/java/com/ralien/erp_system/authn/dao/AuthorityRepository.java
@@ -3,5 +3,5 @@ package com.ralien.erp_system.authn.dao;
 import com.ralien.erp_system.authn.entities.Authority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorityRepository extends JpaRepository<Authority, Long> {
+public interface AuthorityRepository extends JpaRepository<Authority, Integer> {
 }
