+++ src/main/java/com/ralien/erp_system/authn/dao/ClientBranchRepository.java
@@ -3,5 +3,5 @@ package com.ralien.erp_system.authn.dao;
 import com.ralien.erp_system.authn.entities.ClientBranch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientBranchRepository extends JpaRepository<ClientBranch, Long> {
+public interface ClientBranchRepository extends JpaRepository<ClientBranch, Integer> {
 }
